requirements.txt
numpy>=2.1.0
sounddevice==0.4.7
opuslib==3.0.1
Pillow==10.4.0
pystray==0.19.5